Updated the display servers for: 1. Specifying DPI as an 'int' 2. Returning a default DPI in 'headless' mode

+++ chart/org.eclipse.birt.chart.engine/src/org/eclipse/birt/chart/device/IDisplayServer.java
@@ -43,7 +43,7 @@ public interface IDisplayServer
      * 
      * @return
      */
-    double getDpiResolution();
+    int getDpiResolution();
 
     /**
      *
